src/test/java/io/github/vlad324/n792/SolutionTest.java
package io.github.vlad324.n792;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

class SolutionTest {

    private final Solution solution = new Solution();

    private static Stream<Arguments> testDataProvider() {
        return Stream.<Arguments>builder()
            .add(arguments("abcde", new String[]{"a", "bb", "acd", "ace"}, 3))
            .add(arguments("dsahjpjauf", new String[]{"ahjpjau", "ja", "ahbwzgqnuk", "tnmlanowax"}, 2))
            .build();
    }

    @ParameterizedTest
    @MethodSource("testDataProvider")
    void should_produce_expected_result(String s, String[] words, int expected) {
        // when
        final var actual = solution.numMatchingSubseq(s, words);

        // then
        assertThat(actual).isEqualTo(expected);
    }

}